unsplit FileGetter code by os
